+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 * コンフィグのエラーの修正（osとhostのコンフリクト）
 * Linux: 全てのケースに解像度を受け取る様に
 * Linux: 「/proc/cpuinfo」で使ってCPU速さの受け取る様に
-* Kubuntu、Xubuntu、Lubuntuロゴの追加
-* Arcoロゴの追加
+* Kubuntu、Xubuntu、Lubuntu、Arco、Hyperbola、Parabola、とPop OSロゴの追加
+* Ubuntu MATE、Ubuntu Budgie、Ubuntu Lomiri、Ubuntu Unity、とUbuntu Cinnamonロゴの追加
 * Linux: 「hostname」コマンドがなければ、「cat /etc/hostname」を実効する様に
+* Linux: pacman、rpm対応
 * LinuxとBSD: 今から、静的リンクがデフォルトになりました（動的リンクには「make LDFLAGS=-lc」をご利用下さい）
 
 # 0.2.0